requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Cache==0.12
Flask-DebugToolbar==0.8.0
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.1
Jinja2==2.7.1
MarkupSafe==0.18
PIL==1.1.7
PyYAML==3.10
SQLAlchemy==0.8.2
WTForms==1.0.4
Werkzeug==0.9.4
argh==0.23.3
blinker==1.3
coverage==3.6
cracklib==2.8.19
cssselect==0.8
diff-match-patch==20110725.1
gunicorn==18.0
itsdangerous==0.23
lxml==3.2.3
pathtools==0.1.2
psycopg2==2.5.1
pyflakes==0.7.3
pylibmc==1.2.3
pyquery==1.2.4
python-magic==0.4.3
pytidylib==0.2.1
pytz==2013b
simplejson==3.3.0
speaklater==1.3
ujson==1.33
watchdog==0.6.0
wsgiref==0.1.2
